requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.0.7
djangorestframework==3.15.2
dnspython==2.6.1
mongoengine==0.28.2
pycodestyle==2.12.0
pymongo==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
